backend/requirements.txt
# Core Django
Django==4.2.16
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.4.0
django-ratelimit==4.1.0
python-decouple==3.8
setuptools==75.6.0
importlib-metadata==8.5.0

# PDF Processing
pypdf==4.3.1
pdf2image==1.17.0
img2pdf==0.5.1
reportlab==4.2.2
pdfplumber==0.11.4
PyMuPDF==1.24.10

# Document Processing
python-docx==1.1.2
docx2txt==0.8

# Image Processing
Pillow==10.4.0

# Security
cryptography==43.0.1

# Server & Database
gunicorn==23.0.0
psycopg2-binary==2.9.9
dj-database-url==2.2.0
whitenoise==6.7.0